+++ src/Tests/AccountSummary/AccountSummaryData.test.js
@@ -24,7 +24,7 @@ describe('Renders account summary data', () => {
     expect(tr[1] === message).toEqual(true)
   })
   it('Shows Fees', () => {
-    const message = 'Fees'
+    const message = 'Earned from delegate fees'
     let wrapper = mount(<AccountSummaryData {...props} />)
     let tr = wrapper.find('TableCell').map(tr => tr.text())
     expect(tr[3] === message).toEqual(true)
@@ -35,6 +35,7 @@ describe('Renders account summary data', () => {
     let tr = wrapper.find('TableCell').map(tr => tr.text())
     expect(tr[5] === message).toEqual(true)
   })
+  /*
   it('Shows LastClaimRound', () => {
     const message = 'LastClaimRound'
     let wrapper = mount(<AccountSummaryData {...props} />)
@@ -53,4 +54,5 @@ describe('Renders account summary data', () => {
     let tr = wrapper.find('TableCell').map(tr => tr.text())
     expect(tr[11] === message).toEqual(true)
   })
+  */
 })
