+++ src/Tests/Common/UI/Card/CardBody.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Adapter from 'enzyme-adapter-react-16'
 import { configure, mount } from 'enzyme'
-import CardBody from '../../../Components/Common/UI/Card/CardBody'
+import CardBody from '../../../../Components/Common/UI/Card/CardBody'
 
 configure({ adapter: new Adapter() })
